// api/Channel.ts
import { Video } from "@/api/Video";

interface YTImage {
  url: string;

  width: number;
  height: number;
}

export interface Channel {
  type: "channel";
  channelHandle?: string;
  author: string;
  authorId: string;
  authorUrl: string;
  authorBanners: YTImage[];
  authorThumbnails: YTImage[];
  subCount: number;
  totalViews: number;
  joined: number;
  autoGenerated: boolean;
  isFamilyFriendly: boolean;
  description: string;
  descriptionHtml: string;
  allowedRegions: string[];
  tabs: string[];
  authorVerified: boolean;
  latestVideos: Video[];
  relatedChannels: unknown[];
}